cmd/link: make ppc64 __glink_PLTresolve position-independent
This code is only generated when linking cgo internally with fixed
position code. This feature of the internal linker is only supported
on ppc64le/linux targets. This moves ppc64le/linux a little closer
to supporting PIE when internal linking.

This is more similiar to the implementation suggested in the ppc64
elfv2 supplement, and works with both PIE and static code.

Change-Id: I0b64e1c22b9e07b5151378d2ab19ee0e50405fc5