cmd/compile/internal/types2: factor out slice elem computation for copy built-in
Implement singleUnder which determines a single underlying type for a
given type: either the underlying type, or the single underlying type
for a type parameter, if it exists. Use singleUnder instead of optype
for copy built-in.

This CL removes a dependency on optype and also makes the copy built-in
slighty more general for generic arguments (the source argument may be
constrained by a slice or string simultaneously).

Change-Id: Ia329e96afc69a09d2ca3b1f82fe712d4f7ba1d9f