cmd: add support for reading coverage counter data files
DO NOT SUBMIT

Add a set of helper packages for reading counter data files produced
by runs of coverage-instrumented binaries. As a test driver for these
pacakges, this patch also includes a dump tool (for debugging, not
intended for end users) that dumps out the contents of a coverage data
run in human-readable form.

Change-Id: I44167c578f574b4636ab8726e726388531fd3258