go/parser: fix parsing of array or slice constraint types
Now that we allow eliding 'interface' from constraint types, we need to
be a bit more careful about not consuming a '[' when parsing the next
expression after "type T [". We want to check if the next expression is
an identifier (in which case we're in a generic type), but need to
avoid, for example, parsing the primary index expression "_ [N]": index
expressions aren't valid array lengths, so we can instead assume that
this is a type parameter field with array type constraint.

Test cases are added for the related issues #49174 and #49175, along
with a flag to enable tracing error tests.

For #49174
For #49175

Change-Id: I0476ef20c4c134ac537118272f20caaf123ee70e