internal/coverage: add apis for reading/writing counter data
DO NOT SUBMIT

Add utilities for reading and writing of counter data files as part of
the new code coverage implementation.

Currently the contents of the counter data file are emitted in
little-endian form, which for the counters is somewhat painful in that
we need to visit each counter value and properly encode it. It might
be better to instead emit counters in native endianity and then teach
the tools to decode properly in the case of an endianity mismatch.

Change-Id: I124fdcb40fc339a48b64b35264bf24c3be50ddd4