internal/coverage: add apis for reading/writing counter data
Add utilities for reading and writing of counter data files as part of
the new code coverage implementation.

Trivia note: currently the contents of counter data files are
emitted in little-endian form, which for the counters is somewhat
painful in that we need to visit each counter value and properly
encode it. It might be better to instead emit counters in native
endianity and then teach the tools to decode properly in the case of
an endianity mismatch.

Updates #51430.

Change-Id: I124fdcb40fc339a48b64b35264bf24c3be50ddd4