cmd/compile: add test for 49241
The types of the two interfaces should be equal, but they aren't.

It is probably new to generics that two types that are supposed to be
identical can be manufactured in two different packages. I think we'll
need type deduplication in the linker.

Possibly related to CL 356570, but that CL probably isn't the cause.

Update #49241

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: I8a6c70da541c6088a92a01392bc83b61cc130eba