cmd/go,cmd/compile: add a flag to support load address of itab from table
A direct LEAQ is used in every place in the program that needs an itab. But for plugins, maybe there are some dynamically created itabs in the host process. If an iface with dynamically created itab is delivered to the plugin. This itab can't be asserted to the real type of it in the plugin, because the LEAQ always refrence to the itab symbol in the plugin.

I sent CL 351850 one month ago. Cherry suggested that we could make a different object files for plugin, This is the intial support. Add a flag to compiler. When the buildmode is plugin, this flag will be passed to compiler.

For #48532

Change-Id: I0e5b54cd3053cc0dbe3059b24667090d340a8cf2